crypto/easy_md5/requirements.txt
pwn==1.0
pwntools==4.7.0
pycryptodome==3.15.0
